@@ -9,19 +9,17 @@
 #include <insitu/filter.hpp>
 #include "insitu_utils.hpp"
 
-namespace insitu {
-
+namespace insitu
+{
 class FilterFactory
 {
-
 private:
-
     /* nodelet loader does the initialization */
-    nodelet::Loader * nLoader;
+    nodelet::Loader* nLoader;
 
     /* pluginlib loader is used to get a list of loadable insitu filters
        and do the actual loading */
-    pluginlib::ClassLoader<insitu::Filter> * pLoader;
+    pluginlib::ClassLoader<insitu::Filter>* pLoader;
 
     /* variable for create_instance to store into */
     boost::shared_ptr<insitu::Filter> instance_;
@@ -31,31 +29,24 @@ private:
     create_instance(const std::string& lookup_name);
 
 public:
-
     FilterFactory(const std::string& pkg);
 
     ~FilterFactory(void);
 
-    std::vector<std::string>
-    getFilterList(void);
+    std::vector<std::string> getFilterList(void);
 
-    boost::shared_ptr<insitu::Filter>
-    loadFilter(const std::string& filter, const std::string& name);
+    boost::shared_ptr<insitu::Filter> loadFilter(const std::string& filter,
+                                                 const std::string& name);
 
-    bool
-    unloadFilter(const std::string& name);
+    bool unloadFilter(const std::string& name);
 
-    std::string
-    getClassDescription(const std::string& name);
+    std::string getClassDescription(const std::string& name);
 
-    std::string
-    getClassPackage(const std::string& name);
+    std::string getClassPackage(const std::string& name);
 
-    std::string
-    getName(const std::string& name);
-
+    std::string getName(const std::string& name);
 };
 
-} // namespace insitu
+}    // namespace insitu
 
-#endif // insitu_FILTER_FACTORY_HPP
+#endif    // insitu_FILTER_FACTORY_HPP
