@@ -11,40 +11,38 @@
 #include "insitu_utils.hpp"
 #include "filtered_view.hpp"
 
-namespace insitu {
-
+namespace insitu
+{
 class ModeContainer : public QWidget
 {
-
-Q_OBJECT
+    Q_OBJECT
 private:
     // UI elements
-    QPushButton * addViewButton;
-    QPushButton * tileButton;
-    QPushButton * cascadeButton;
-    QMdiArea * container;
+    QPushButton* addViewButton;
+    QPushButton* tileButton;
+    QPushButton* cascadeButton;
+    QMdiArea* container;
 
     // layout element
-    QGridLayout * layout;
+    QGridLayout* layout;
 
     // ROS
-    ros::NodeHandle * nh;
+    ros::NodeHandle* nh;
 
 public Q_SLOTS:
     void tile(void);
     void cascade(void);
 
 public:
-    ModeContainer(QString name, QWidget * parent = nullptr);
+    ModeContainer(QString name, QWidget* parent = nullptr);
 
     ~ModeContainer(void);
 
-    void addView(FilteredView * view);
+    void addView(FilteredView* view);
 
     const ros::NodeHandle& getNodeHandle(void);
-
 };
 
-} // end namespace insitu
+}    // end namespace insitu
 
 #endif
