@@ -17,21 +17,23 @@
 using namespace std::chrono_literals;
 
 template <typename TF, typename TDuration, class... TArgs>
-std::result_of_t<TF&&(TArgs&&...)> run_with_timeout(TF&& f, TDuration timeout, TArgs&&... args)
+std::result_of_t<TF && (TArgs && ...)> run_with_timeout(TF&& f,
+                                                        TDuration timeout,
+                                                        TArgs&&... args)
 {
-    using R = std::result_of_t<TF&&(TArgs&&...)>;
+    using R = std::result_of_t<TF && (TArgs && ...)>;
     std::packaged_task<R(TArgs...)> task(f);
     auto future = task.get_future();
     std::thread thr(std::move(task), std::forward<TArgs>(args)...);
     if (future.wait_for(timeout) != std::future_status::timeout)
     {
-       thr.join();
-       return future.get(); // this will propagate exception from f() if any
+        thr.join();
+        return future.get();    // this will propagate exception from f() if any
     }
     else
     {
-       thr.detach(); // we leave the thread still running, non-ideal
-       throw std::runtime_error("Timeout");
+        thr.detach();    // we leave the thread still running, non-ideal
+        throw std::runtime_error("Timeout");
     }
 }
 
@@ -42,10 +44,10 @@ void try_run_with_timeout(TF&& f, TDuration timeout, TArgs&&... args)
     {
         run_with_timeout(f, timeout, std::forward<TArgs>(args)...);
     }
-    catch (std::exception & e)
+    catch (std::exception& e)
     {
         std::cout << "Exception: " << e.what() << std::endl;
     }
 }
 
-#endif // insitu_TIMED_RUN_HPP
+#endif    // insitu_TIMED_RUN_HPP
