@@ -1,10 +1,11 @@
 #include "filter_factory.hpp"
 
-namespace insitu {
-
+namespace insitu
+{
 FilterFactory::FilterFactory(const std::string& pkg)
 {
-    nLoader = new nodelet::Loader(boost::bind(&FilterFactory::create_instance, this, _1));
+    nLoader = new nodelet::Loader(
+        boost::bind(&FilterFactory::create_instance, this, _1));
     pLoader = new pluginlib::ClassLoader<insitu::Filter>(pkg, "insitu::Filter");
 }
 
@@ -14,8 +15,7 @@ FilterFactory::~FilterFactory(void)
     if (pLoader) delete pLoader;
 }
 
-std::vector<std::string>
-FilterFactory::getFilterList(void)
+std::vector<std::string> FilterFactory::getFilterList(void)
 {
     return pLoader->getDeclaredClasses();
 }
@@ -31,9 +31,12 @@ FilterFactory::loadFilter(const std::string& filter, const std::string& name)
     /* this will call create_instance and set instance_ */
     bool loaded = nLoader->load(name, filter, rmap_, argv_);
 
-    if (loaded) {
+    if (loaded)
+    {
         // TODO track
-    } else {
+    }
+    else
+    {
         // TODO error
         throw std::runtime_error("Failed to load Filter: " + filter);
     }
@@ -51,28 +54,24 @@ FilterFactory::create_instance(const std::string& lookup_name)
     return instance_;
 }
 
-bool
-FilterFactory::unloadFilter(const std::string& name)
+bool FilterFactory::unloadFilter(const std::string& name)
 {
     return nLoader->unload(name);
 }
 
-std::string
-FilterFactory::getClassDescription(const std::string& name)
+std::string FilterFactory::getClassDescription(const std::string& name)
 {
     return pLoader->getClassDescription(name);
 }
 
-std::string
-FilterFactory::getClassPackage(const std::string& name)
+std::string FilterFactory::getClassPackage(const std::string& name)
 {
     return pLoader->getClassPackage(name);
 }
 
-std::string
-FilterFactory::getName(const std::string& name)
+std::string FilterFactory::getName(const std::string& name)
 {
     return pLoader->getName(name);
 }
 
-}
+}    // namespace insitu
